refactor : split query and mutations
